add replace_all
Should allow a 'replace_all' keyword...
Why this wasn't implemented: probably because replacing all contents creates an unknown html state.  When used without forethought, it can replace significant portions of html.

Probably needs a validator to ensure replace_all is only used with a #target, and perhaps on an element with few or no children.  But this should do for testing.